i2d

+++ lib/src/models/track_meta.dart
@@ -87,5 +87,5 @@ class TrackMeta extends Object {
 
   ///
   @JsonKey(name: 'rhythm_version')
-  int? rhythmVersion;
+  double? rhythmVersion;
 }
